+++ ChangeLog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IrrepsArray.astype` to cast the underlying array
 - `e3nn.flax.MultiLayerPerceptron` and `e3nn.haiku.MultiLayerPerceptron`
 
+### Fixed
+- set dtype for all `jnp.zeros(..., dtype)` calls in the codebase
+
 ### Removed
 - `e3nn.full_tensor_product` in favor of `e3nn.tensor_product`
 - `e3nn.FunctionalTensorSquare` in favor of `e3nn.tensor_square`