+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.14.0] - 2022-12-16
 ### Added
 - `IrrepsArray.astype` to cast the underlying array
 - `e3nn.flax.MultiLayerPerceptron` and `e3nn.haiku.MultiLayerPerceptron`